+++ CHANGELOG.md
@@ -3,6 +3,11 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
+## [1.8.8] - 2021-11-04
+## Changes
+- fix some connect/disconnect case
+- fix some remote command not send 
+
 ## [1.8.7] - 2021-11-01
 ## Changes
 - performance improvement